Fix upload issue in 2.2.5 by adding file and mime types
Magento 2.2.5 adds a mime type check in the uploadFile method in Magento/Cms/Model/Wysiwyg/Images/Storage.php.

If the mime type check fails the upload is aborted, giving the error message ”File validation failed.”

The allowed mime types are defined in Magento/Cms/etc/di.xml

So, we need to add file and mime types we want to support to the allowed and image_allowed arrays in di.xml.